% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitais.R
\name{capitais}
\alias{capitais}
\title{Current weather conditions of capitals}
\usage{
capitais()
}
\value{
The function returns a data frame
}
\description{
This function allows the user to retrieve data from
CPTEC/INPE current weather conditions of capitals.
The data frame returned contain actual temperature, 
atmospheric pressure, humidity, direction and wind
intensity
}
\examples{
\dontrun{
capitais()
}

}
\seealso{
\code{\link{prev4dias}}
}
\author{
Renato Prado Siqueira \email{<rpradosiqueira@gmail.com>}
}
\keyword{brazil}
\keyword{forecast}
\keyword{weather}
