% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condicoes_tempo.R
\name{condicoes_tempo}
\alias{condicoes_tempo}
\title{Data frame containing "tempo" descriptions}
\usage{
condicoes_tempo()
}
\value{
The function returns a sigle data frame.
}
\description{
Data frame containing "tempo" descriptions
}
\examples{

\dontrun{

condicoes_tempo()

}

}
\seealso{
\code{\link{estacoes_metar}}
}
\author{
Renato Prado Siqueira \email{rpradosiqueira@gmail.com}
}
\keyword{brazil}
\keyword{forecast}
\keyword{weather}
