% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estacoes_metar.R
\name{estacoes_metar}
\alias{estacoes_metar}
\title{Data frame containing "Metar" initials and descriptions}
\usage{
estacoes_metar()
}
\value{
The function returns a sigle data frame.
}
\description{
Data frame containing "Metar" initials and descriptions
}
\examples{
\dontrun{
estacoes_metar()
}

}
\seealso{
\code{\link{condicoes_tempo}}
}
\author{
Renato Prado Siqueira \email{<rpradosiqueira@gmail.com>}
}
\keyword{brazil}
\keyword{forecast}
\keyword{weather}
