% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estendidaLatLon.R
\name{estendidaLatLon}
\alias{estendidaLatLon}
\title{Retrieve 14 days weather forecast data from CPTEC/INPE}
\usage{
estendidaLatLon(lat, lon)
}
\arguments{
\item{lat}{The latitude numeric/character vector}

\item{lon}{The longitude numeric/character vector}
}
\value{
The function returns a data frame printed by the parameters input
}
\description{
This function allows the user to retrieve data from
CPTEC/INPE weather forecast API for the next 14 days
informing latitude and longitude of the local.
The data frame returned contain maximum and minimum 
temperature, ultra violet index and weather condition.
}
\examples{
\dontrun{
## Requesting weather forecast with a latitude/longitude information
prevLatLon("-20.432257", "-54.569751")
}

}
\seealso{
\code{\link{prevEstendida}}
}
\author{
Renato Prado Siqueira \email{<rpradosiqueira@gmail.com>}
}
\keyword{brazil}
\keyword{forecast}
\keyword{weather}
