% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iuv.R
\name{iuv}
\alias{iuv}
\title{Get current ultra violet index}
\usage{
iuv(id)
}
\arguments{
\item{id}{The city's ID}
}
\value{
The function returns a data frame printed by city id
}
\description{
This function allows the user to retrieve the last updated 
data from CPTEC/INPE ultra violet index for a determined city
}
\examples{
\dontrun{
## Requesting the last updated of the ultra violet index drom Campo Grande/MS
iuv(225)
}

}
\seealso{
\code{\link{prevEstendida}}
}
\author{
Renato Prado Siqueira \email{<rpradosiqueira@gmail.com>}
}
\keyword{brazil}
\keyword{forecast}
\keyword{index}
\keyword{ultra}
\keyword{violet}
\keyword{weather}
