% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listaCidades.R
\name{listaCidades}
\alias{listaCidades}
\title{Search for cities ID}
\usage{
listaCidades(x)
}
\arguments{
\item{x}{A single character to search.}
}
\value{
The function returns a data frame with the results of the search.
}
\description{
This function allows the user to discover which are
the city's ID of interest.
}
\examples{
\dontrun{
## Looking for the city of Campo Grande
listaCidades("campo grande")
}

}
\seealso{
\code{\link{prev4dias}}
}
\author{
Renato Prado Siqueira \email{<rpradosiqueira@gmail.com>}
}
\keyword{brazil}
\keyword{forecast}
\keyword{weather}
