% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metar.R
\name{metar}
\alias{metar}
\title{Retrieve weather condition data from Metars}
\usage{
metar(x)
}
\arguments{
\item{x}{A character vector with the desired weather station(s) code(s)}
}
\value{
The function returns a data frame
}
\description{
This function allows the user to retrieve data from
airport surface stations (Metar), coded and made
available by the meteorological network of the
Aeronautics Command - REDEMET
}
\examples{
\dontrun{
## Requesting weather condition from SBGR and SBTK stations
metar(c("SBGR", "SBTK"))
}

}
\seealso{
\code{\link{estacoes_metar}}
}
\author{
Renato Prado Siqueira \email{<rpradosiqueira@gmail.com>}
}
\keyword{brazil}
\keyword{forecast}
\keyword{weather}
