% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev4dias.R
\name{prev4dias}
\alias{prev4dias}
\title{Retrieve 4 days weather forecast data from CPTEC/INPE}
\usage{
prev4dias(id)
}
\arguments{
\item{id}{The city's ID}
}
\value{
The function returns a data frame printed by city id
}
\description{
This function allows the user to retrieve data from
CPTEC/INPE weather forecast API for the next 4 days.
The data frame returned contain maximum and minimum
temperature, ultra violet index and weather condition
(tempo).
}
\examples{
## Requesting weather forecast from the city of Campo Grande/MS
prev4dias(225)

## Requesting weather forecast from the city of Rio de Janeiro/RJ
prev4dias(241)

}
\seealso{
\code{\link{prev7dias}}
}
\author{
Renato Prado Siqueira \email{rpradosiqueira@gmail.com}
}
\keyword{brazil}
\keyword{forecast}
\keyword{weather}
