% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev7dias.R
\name{prev7dias}
\alias{prev7dias}
\title{Retrieve 7 days weather forecast data from CPTEC/INPE}
\usage{
prev7dias(id)
}
\arguments{
\item{id}{The city's ID}
}
\value{
The function returns a data frame printed by city id
}
\description{
This function allows the user to retrieve data from
CPTEC/INPE weather forecast API for the next 7 days.
The data frame returned contain maximum and minimum
temperature, ultra violet index and weather condition.
}
\examples{
## Requesting weather forecast from the city of Campo Grande/MS
prev7dias(225)

}
\seealso{
\code{\link{prev7dias}}
}
\author{
Renato Prado Siqueira \email{<rpradosiqueira@gmail.com>}
}
\keyword{brazil}
\keyword{forecast}
\keyword{weather}
