% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevEstendida.R
\name{prevEstendida}
\alias{prevEstendida}
\title{Retrieve 14 days weather forecast data from CPTEC/INPE}
\usage{
prevEstendida(id)
}
\arguments{
\item{id}{The city's ID}
}
\value{
The function returns a data frame printed by city id
}
\description{
This function allows the user to retrieve data from
CPTEC/INPE weather forecast API for the next 14 days.
The data frame returned contain maximum and minimum 
temperature and weather condition (tempo).
}
\examples{
\dontrun{
## Requesting weather forecast from the city of Campo Grande/MS
prevEstendida(225)
}

}
\seealso{
\code{\link{prev7dias}}
}
\author{
Renato Prado Siqueira \email{<rpradosiqueira@gmail.com>}
}
\keyword{brazil}
\keyword{forecast}
\keyword{weather}
